Replace MiqSshUtil with MiqSshUtilV2
MiqSshUtil is not used as a script and there is no chance we are
using a net-ssh version < 2 at this point.
